REDU-466; Create Likert custom component for likert type repeating group
