+++ src/altinn-app-frontend/src/components/base/RadioButtons/RadioButtonsContainerComponent.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { IComponentProps } from '../../index';
+import type { IMapping, IOption, IOptionSource } from 'src/types';
+import { LayoutStyle } from 'src/types';
+import { useRadioButtons } from 'src/components/base/RadioButtons/radioButtonsUtils';
+import { ControlledRadioGroup } from 'src/components/base/RadioButtons/ControlledRadioGroup';
+
+export interface IRadioButtonsContainerProps extends IComponentProps {
+  validationMessages?: any;
+  options?: IOption[];
+  optionsId?: string;
+  preselectedOptionIndex: number;
+  title: string;
+  mapping?: IMapping;
+  layout?: LayoutStyle;
+  source?: IOptionSource;
+}
+
+export const RadioButtonContainerComponent = (
+  props: IRadioButtonsContainerProps,
+) => {
+  const useRadioProps = useRadioButtons(props);
+  return <ControlledRadioGroup {...props} {...useRadioProps} />;
+};
