+++ src/altinn-app-frontend/src/components/index.ts
@@ -9,7 +9,7 @@ import { FileUploadWithTagComponent } from './base/FileUpload/FileUploadWithTag/
 import { HeaderComponent } from './base/HeaderComponent';
 import { InputComponent } from './base/InputComponent';
 import { ParagraphComponent } from './base/ParagraphComponent';
-import { RadioButtonContainerComponent } from './base/RadioButtonsContainerComponent';
+import { RadioButtonContainerComponent } from './base/RadioButtons/RadioButtonsContainerComponent';
 import { TextAreaComponent } from './base/TextAreaComponent';
 import { ImageComponent } from './base/ImageComponent';
 import { NavigationButtons as NavigationButtonsComponent } from './presentation/NavigationButtons';
@@ -19,6 +19,7 @@ import { IGenericComponentProps } from './GenericComponent';
 import { IComponentFormData } from 'src/utils/formComponentUtils';
 import { ILanguage } from 'altinn-shared/types';
 import type { ITextResourceBindings } from 'src/types';
+import { LikertComponent } from 'src/components/base/LikertComponent';
 
 export interface IComponent {
   name: string;
@@ -48,6 +49,7 @@ export enum ComponentTypes {
   InstantiationButton,
   AttachmentList,
   NavigationBar,
+  Likert,
 }
 
 export const textComponents: IComponent[] = [
@@ -160,6 +162,16 @@ export const schemaComponents: IComponent[] = [
     Tag: NavigationBar,
     Type: ComponentTypes.NavigationBar,
   },
+  {
+    name: 'Likert',
+    Tag: LikertComponent,
+    Type: ComponentTypes.Likert,
+    customProperties: {
+      options: [],
+      required: false,
+      readOnly: false,
+    },
+  },
 ];
 
 export const advancedComponents: IComponent[] = [
@@ -175,7 +187,11 @@ export const advancedComponents: IComponent[] = [
 ];
 
 export interface IComponentProps extends IGenericComponentProps {
-  handleDataChange: (value: string, key?: string, skipValidation?: boolean) => void;
+  handleDataChange: (
+    value: string,
+    key?: string,
+    skipValidation?: boolean,
+  ) => void;
   handleFocusUpdate: (componentId: string, step?: number) => void;
   getTextResource: (key: string) => React.ReactNode;
   getTextResourceAsString: (key: string) => string;
